Rename src/k8s-engine/ -> src/backend/
The backends handle more the k8s.

Note that the interfaces (and the imported names) are not changed yet; that
will come in later PRs.
