+++ background.ts
@@ -6,15 +6,15 @@ import util from 'util';
 import Electron from 'electron';
 import _ from 'lodash';
 
+import K8sFactory from '@/backend/factory';
+import { getImageProcessor } from '@/backend/images/imageFactory';
+import { ImageProcessor } from '@/backend/images/imageProcessor';
+import * as K8s from '@/backend/k8s';
+import { Steve } from '@/backend/steve';
 import * as settings from '@/config/settings';
 import { IntegrationManager, getIntegrationManager } from '@/integrations/integrationManager';
 import { removeLegacySymlinks, PermissionError } from '@/integrations/legacy';
 import { getPathManagerFor, PathManagementStrategy, PathManager } from '@/integrations/pathManager';
-import K8sFactory from '@/k8s-engine/factory';
-import { getImageProcessor } from '@/k8s-engine/images/imageFactory';
-import { ImageProcessor } from '@/k8s-engine/images/imageProcessor';
-import * as K8s from '@/k8s-engine/k8s';
-import { Steve } from '@/k8s-engine/steve';
 import { CommandWorkerInterface, HttpCommandServer } from '@/main/commandServer/httpCommandServer';
 import SettingsValidator from '@/main/commandServer/settingsValidator';
 import { HttpCredentialHelperServer } from '@/main/credentialServer/httpCredentialHelperServer';
