+++ src/backend/progressTracker.ts
