+++ src/components/PortForwarding.vue
@@ -85,10 +85,10 @@
 <script lang="ts">
 import Vue from 'vue';
 
+import * as K8s from '@/backend/k8s';
 import Banner from '@/components/Banner.vue';
 import SortableTable from '@/components/SortableTable/index.vue';
 import Checkbox from '@/components/form/Checkbox.vue';
-import * as K8s from '@/k8s-engine/k8s';
 
 import type { PropType } from 'vue';
