+++ src/pages/PortForwarding.vue
@@ -24,9 +24,9 @@
 import { ipcRenderer } from 'electron';
 import Vue from 'vue';
 
+import * as K8s from '@/backend/k8s';
 import PortForwarding from '@/components/PortForwarding.vue';
 import { defaultSettings, Settings } from '@/config/settings';
-import * as K8s from '@/k8s-engine/k8s';
 
 export default Vue.extend({
   components: { PortForwarding },
